Fix spelling mistake
The following originally aimed to fix the "sofware" -> "software"
spelling mistake. To prevent this happening again, tests where added
to verify this and in doing so, also spotted other errors with in the
origin file. Additional tests where then added to make sure that those
ones also don't break in the future.

Summary:

 - software spelling fix
 - OriginEnterpriseID error not being returned
 - OriginEnterpriseID path returned partial or missing items in the path